api/requirements.txt
bcrypt==5.0.0
bidict==0.23.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dataclasses-json==0.5.7
defusedxml==0.7.1
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-SocketIO==5.6.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.62.1
fpdf==1.7.2
fpdf2==2.8.7
gevent==24.1.1
gevent-websocket==0.10.1
greenlet==3.3.2
gunicorn==21.2.0
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
libcst==1.8.6
MarkupSafe==3.0.3
marshmallow==3.26.2
marshmallow-enum==1.5.1
mypy_extensions==1.1.0
openpyxl==3.1.5
packaging==26.0
pillow==12.1.1
psutil==7.2.2
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.12.1
pyre-extensions==0.0.32
python-dotenv==1.0.0
python-engineio==4.13.1
python-socketio==5.16.1
PyYAML==6.0.3
simple-websocket==1.1.0
SQLAlchemy==2.0.48
tabulate==0.10.0
TestSlide==2.7.1
tomli==2.4.0
tomli_w==1.2.0
typeguard==2.13.3
werkzeug==3.0.1
typing-inspect==0.9.0
typing_extensions==4.15.0
Werkzeug==3.0.1
wsproto==1.3.2
WTForms==3.2.1
